feat(save): implement extension system for declarative save mapping and serialization
Introduce a comprehensive extension system for the save module to simplify
serialization of complex types through declarative attributes and type-aware
mapping contracts. This enables Quality of Life improvements for developers
working with the save system.

Changes include:
- Add SaveMappingService for automatic type contract generation and caching
- Implement SaveTypeContract and SaveMemberContract for metadata-driven
  serialization
- Add declarative attributes (SaveField, SaveCollection, SaveIgnore, SaveObject)
  for fine-grained control over serialization behavior
- Implement SaveReadContext and SaveWriteContext for stateful serialization
  operations
- Add ScalarFieldIO for primitive type handling
- Extend SaveWriterService with fluent extension methods for improved API
- Reorganize tests with improved integration test coverage

+++ Source/Nomad.OnlineServices.Steam/Private/Services/SteamService.cs
@@ -217,7 +217,7 @@ namespace Nomad.OnlineServices.Steam.Private.Services
 			SteamAPI.RunCallbacks();
 
 			_inputService?.Frame(
-				Stopwatch.GetTimestamp() * 1000.0f / Stopwatch.Frequency
+				(long)( Stopwatch.GetTimestamp() * 1000.0f / Stopwatch.Frequency )
 			);
 			_statsService.StoreStats();
 		}
